Link state to hash
